favorites changing color only when specific item is clicked, furnitureId is undefined
