+++ src/redux/reducers/furnitureType.reducer.js
@@ -1,7 +1,7 @@
 const furnitureTypeReducer = (state = [], action) => {
     switch (action.type) {
         case 'SET_FURNITURE_TYPE':
-            console.log('reducer', action.payload);
+            // console.log('reducer', action.payload);
             return action.payload;
         default:
             return state;
