+++ src/redux/reducers/myFavorites.reducer.js
@@ -1,6 +1,7 @@
 const myFavorites = (state=[], action) => {
     switch (action.type) {
         case 'SET_MY_FAVORITES':
+            console.log('myFavorites:', action.payload);
             return action.payload;
         default:
             return state;
